+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -13,6 +13,10 @@
 * Added `ByPropertyIdGrouper`
 * Added `BestStatementsFinder`
 
+#### Deprecations
+
+* Deprecated `ByPropertyIdArray` in favour of `ByPropertyIdGrouper`
+
 ## Version 1.0 (2014-09-02)
 
 #### Breaking changes